moved FirearmNotes to own folder
